refactor and readd attribute tool
