+++ src/beaker_climate/beaker_climate/procedures/python3/get_attribute_values.py
@@ -1,0 +1,1 @@
+catalog.get_available_values("{{attribute}}")
